refactor: move utility modules to utils
