--- src/decNumToHex.ts
@@ -1,8 +1,0 @@
-export function decNumToHex (decNum: number): string {
-  decNum = Math.floor(decNum);
-  if (isNaN(decNum)) {
-    return '00';
-  }
-
-  return ('0' + decNum.toString(16)).slice(-2);
-}
