--- src/hexNumToDec.ts
@@ -1,7 +1,0 @@
-export function hexNumToDec (hexNum: string): number {
-  if (isNaN(parseInt(hexNum, 16))) {
-    return 0;
-  }
-
-  return parseInt(hexNum, 16);
-}
