+++ src/hexToRgb.ts
@@ -1,6 +1,6 @@
-import { hexNumToDec } from './hexNumToDec';
 import { parseHex } from './parseHex';
 import { HexColor, RgbColor } from './types';
+import { hexNumToDec } from './utils';
 
 export function hexToRgb (hex: HexColor | string | null) {
   if (!hex) {
