+++ src/index.ts
@@ -1,80 +1,36 @@
-import { cmykToHex } from './cmykToHex';
-import { cmykToHsl } from './cmykToHsl';
-import { cmykToRgb } from './cmykToRgb';
-import { decNumToHex } from './decNumToHex';
-import { expandHexShorthand } from './expandHexShorthand';
-import { getColorModel } from './getColorModel';
-import { getRandomColor } from './getRandomColor';
-import { hexNumToDec } from './hexNumToDec';
-import { hexToCmyk } from './hexToCmyk';
-import { hexToHsl } from './hexToHsl';
-import { hexToHsv } from './hexToHsv';
-import { hexToRgb } from './hexToRgb';
-import { hslToCmyk } from './hslToCmyk';
-import { hslToHex } from './hslToHex';
-import { hslToHsv } from './hslToHsv';
-import { hslToRgb } from './hslToRgb';
-import { hsvToHex } from './hsvToHex';
-import { hsvToHsl } from './hsvToHsl';
-import { hsvToRgb } from './hsvToRgb';
-import { isHexShorthand } from './isHexShorthand';
-import { isValidColor } from './isValidColor';
-import { mixColors } from './mixColors';
-import { normalizeDecNum } from './normalizeDecNum';
-import { parseCmyk } from './parseCmyk';
-import { parseHex } from './parseHex';
-import { parseHsl } from './parseHsl';
-import { parseHsv } from './parseHsv';
-import { parseRgb } from './parseRgb';
-import { rgbToCmyk } from './rgbToCmyk';
-import { rgbToHex } from './rgbToHex';
-import { rgbToHsl } from './rgbToHsl';
-import { rgbToHsv } from './rgbToHsv';
-import { toCmyk } from './toCmyk';
-import { toHex } from './toHex';
-import { toHsl } from './toHsl';
-import { toHsv } from './toHsv';
-import { toRgb } from './toRgb';
-import * as Colors from './types';
+export { cmykToHex } from './cmykToHex';
+export { cmykToHsl } from './cmykToHsl';
+export { cmykToRgb } from './cmykToRgb';
+export { expandHexShorthand } from './expandHexShorthand';
+export { whichModel } from './whichModel';
+export { hexToCmyk } from './hexToCmyk';
+export { hexToHsl } from './hexToHsl';
+export { hexToHsv } from './hexToHsv';
+export { hexToRgb } from './hexToRgb';
+export { hslToCmyk } from './hslToCmyk';
+export { hslToHex } from './hslToHex';
+export { hslToHsv } from './hslToHsv';
+export { hslToRgb } from './hslToRgb';
+export { hsvToHex } from './hsvToHex';
+export { hsvToHsl } from './hsvToHsl';
+export { hsvToRgb } from './hsvToRgb';
+export { isHexShorthand } from './isHexShorthand';
+export { isValidColor } from './isValidColor';
+export { mixColors } from './mixColors';
+export { parseCmyk } from './parseCmyk';
+export { parseHex } from './parseHex';
+export { parseHsl } from './parseHsl';
+export { parseHsv } from './parseHsv';
+export { parseRgb } from './parseRgb';
+export { rgbToCmyk } from './rgbToCmyk';
+export { rgbToHex } from './rgbToHex';
+export { rgbToHsl } from './rgbToHsl';
+export { rgbToHsv } from './rgbToHsv';
+export { toCmyk } from './toCmyk';
+export { toHex } from './toHex';
+export { toHsl } from './toHsl';
+export { toHsv } from './toHsv';
+export { toRgb } from './toRgb';
+export * from './types';
 
-export default {
-  Colors,
-  cmykToHex,
-  cmykToHsl,
-  cmykToRgb,
-  decNumToHex,
-  expandHexShorthand,
-  getColorModel,
-  getRandomColor,
-  hexNumToDec,
-  hexToCmyk,
-  hexToHsl,
-  hexToHsv,
-  hexToRgb,
-  hslToCmyk,
-  hslToHex,
-  hslToHsv,
-  hslToRgb,
-  hsvToHex,
-  hsvToHsl,
-  hsvToRgb,
-  isHexShorthand,
-  isValidColor,
-  mixColors,
-  normalizeDecNum,
-  parseCmyk,
-  parseHex,
-  parseHsl,
-  parseHsv,
-  parseRgb,
-  rgbToCmyk,
-  rgbToHex,
-  rgbToHsl,
-  rgbToHsv,
-  toCmyk,
-  toHex,
-  toHsl,
-  toHsv,
-  toRgb,
-  version: '__VERSION__'
-};
+export const version = '__VERSION__';
