+++ src/isValidColor.ts
@@ -1,10 +1,9 @@
-import { getColorModel } from './getColorModel';
 import { parseCmyk } from './parseCmyk';
 import { parseHex } from './parseHex';
 import { parseHsl } from './parseHsl';
 import { parseHsv } from './parseHsv';
 import { parseRgb } from './parseRgb';
-import { Color } from './types';
+import { whichModel } from './whichModel';
 
 /**
  * Checks if the given color string is valid (parsable).
@@ -12,7 +11,7 @@ import { Color } from './types';
  * @param {String} color The color string to be checked.
  */
 export function isValidColor (color: string | null): boolean {
-  const model = getColorModel(color);
+  const model = whichModel(color);
 
   if (model === 'rgb') {
     return !parseRgb(color).invalid;
